add flake8 tests and rename fixture
